# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Recipe Book API",
        "description": "API for managing recipes, cuisines, ingredients, and user authentication",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/recipes": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Get list of recipes",
                "description": "Get list of recipes",
                "operationId": "e6ebf01e62f938e7ce367c7b2fc94110",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Filter by recipe title (case-insensitive search)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Soup"
                        }
                    },
                    {
                        "name": "cuisine_id",
                        "in": "query",
                        "description": "Filter by cuisine ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of recipes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Recipe"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "session": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Create a new recipe",
                "description": "Create a new recipe",
                "operationId": "5ed3009a1c8ce5b784cac78533de2c73",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "cuisine_id",
                                    "steps"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Carrot Soup"
                                    },
                                    "cuisine_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A delicious soup"
                                    },
                                    "prep_time": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "cook_time": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "servings": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "steps": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "step_number": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "order": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Chop vegetables"
                                                },
                                                "ingredients": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "ingredient_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "amount": {
                                                                "type": "number",
                                                                "example": 500
                                                            },
                                                            "unit": {
                                                                "type": "string",
                                                                "example": "g"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recipe created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recipe created successfully"
                                        },
                                        "recipe": {
                                            "$ref": "#/components/schemas/Recipe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "session": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Recipe": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Carrot Soup"
                    },
                    "cuisine_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "description": {
                        "type": "string",
                        "example": "A delicious soup"
                    },
                    "prep_time": {
                        "type": "integer",
                        "example": 15,
                        "nullable": true
                    },
                    "cook_time": {
                        "type": "integer",
                        "example": 30,
                        "nullable": true
                    },
                    "servings": {
                        "type": "integer",
                        "example": 4,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "cuisine": {
                        "$ref": "#/components/schemas/Cuisine"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "steps": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RecipeStep"
                        }
                    },
                    "attachment": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Attachment"
                        }
                    }
                },
                "type": "object"
            },
            "Cuisine": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Italian"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "RecipeStep": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "recipe_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "step_number": {
                        "type": "integer",
                        "example": 1
                    },
                    "description": {
                        "type": "string",
                        "example": "Chop vegetables"
                    },
                    "order": {
                        "type": "integer",
                        "example": 0
                    },
                    "ingredients": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/IngredientWithAmount"
                        }
                    }
                },
                "type": "object"
            },
            "IngredientWithAmount": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Carrot"
                    },
                    "pivot": {
                        "properties": {
                            "amount": {
                                "type": "number",
                                "example": 500
                            },
                            "unit": {
                                "type": "string",
                                "example": "g"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Attachment": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "url": {
                        "type": "string",
                        "example": "http://localhost/storage/attachments/image.jpg"
                    },
                    "name": {
                        "type": "string",
                        "example": "image.jpg"
                    },
                    "original_name": {
                        "type": "string",
                        "example": "photo.jpg"
                    },
                    "mime": {
                        "type": "string",
                        "example": "image/jpeg"
                    },
                    "size": {
                        "type": "integer",
                        "example": 102400
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "session": {
                "type": "apiKey",
                "description": "Session-based authentication using Laravel sessions",
                "name": "laravel_session",
                "in": "cookie"
            }
        }
    },
...
}